Mv connection_url building
