+++ deliverydb/db.py
@@ -3,7 +3,6 @@ import logging
 import sqlalchemy
 
 from deliverydb.model import Base, Scan
-import model.compliancedb
 
 
 class DeliveryDB:
@@ -37,22 +36,3 @@ class DeliveryDB:
         )
 
         self.Session.add(scan)
-
-
-def database_connection_url_from_cfg(
-    db_cfg: model.compliancedb.ComplianceDbConfig,
-    dialect: str = 'postgresql',
-    overwrite_hostname: str = None,
-    overwrite_username: str = None,
-    overwrite_password: str = None,
-    overwrite_port: int = None,
-) -> str:
-    # use overwrite values if present
-    username = overwrite_username or db_cfg.credentials().username()
-    password = overwrite_password or db_cfg.credentials().password()
-    hostname = overwrite_hostname or db_cfg.hostname()
-    port = overwrite_port or db_cfg.port()
-
-    # Potentially, this function could be used to create connection urls
-    # to different types of databases, therefore the "dialect" can vary.
-    return f'{dialect}://{username}:{password}@{hostname}:{port}'
